Add files via upload
The project is based on the same concept of Blockchain where i have developed a Cryptocurrency using Python with Flask Framework. The project is GUI based where a sender can generate his/her Private Key as well as Public key.
The functionality of the client will be encapsulated on the Port Server availabe on the Machine.For Bob, the Port Server will run on Port 8082 and for Alice the Port Server will be running on port 8081.For nodes,we are having three Web Servers ie. Node1, Node2,and Node3. Blockchain.py is for Nodes and blockchain_client.py is for Clients.
Both Bob as well as Alice can talk with any nodes.Python Flask Framework has been used.